+++ tests/test_screens.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import asyncio
 import sys
 import threading
@@ -5,7 +7,7 @@ import threading
 import pytest
 
 from textual import work
-from textual.app import App, ComposeResult, ScreenError, ScreenStackError
+from textual.app import App, ComposeResult, ScreenStackError
 from textual.events import MouseMove
 from textual.geometry import Offset
 from textual.screen import Screen
@@ -301,8 +303,10 @@ async def test_dismiss_non_top_screen():
     app = MyApp()
     async with app.run_test() as pilot:
         await pilot.press("p")
-        with pytest.raises(ScreenError):
-            await app.bottom.dismiss()
+        # A noop if not the top
+        stack = list(app.screen_stack)
+        await app.bottom.dismiss()
+        assert app.screen_stack == stack
 
 
 async def test_dismiss_action():
@@ -322,6 +326,25 @@ async def test_dismiss_action():
     async with app.run_test() as pilot:
         await pilot.press("y")
         assert app.bingo
+
+
+async def test_dismiss_action_no_argument():
+    class ConfirmScreen(Screen[bool]):
+        BINDINGS = [("y", "dismiss", "Dismiss")]
+
+    class MyApp(App[None]):
+        bingo = False
+
+        def on_mount(self) -> None:
+            self.push_screen(ConfirmScreen(), callback=self.callback)
+
+        def callback(self, result: bool | None) -> None:
+            self.bingo = result
+
+    app = MyApp()
+    async with app.run_test() as pilot:
+        await pilot.press("y")
+        assert app.bingo is None
 
 
 async def test_switch_screen_no_op():